Changed cookie auth settings to work on localhost, also cors
Changed cors to localhost:5173 and session cookies to work cross-origin
and to not be session cookies so that the broser does not delete them so
often.

This is a temporary setting that will likely change in the future on
deploy and that I can later put on some env settings.
It would be interesting to create a DEV mode where i set these to work
on localhost and a PROD setting for production envs